Float32

+++ models/demos/yolov11m/tt/model_preprocessing.py
@@ -68,7 +68,7 @@ def create_yolov11_input_tensors(
         
         ttnn_input_tensor = ttnn.from_torch(
             torch_input_tensor,
-            dtype=ttnn.bfloat16,
+            dtype=ttnn.float32,  # Changed from bfloat16 to preserve input diversity
             layout=ttnn.ROW_MAJOR_LAYOUT,
             device=device,
             memory_config=input_mem_config,
